(#20) Upgrade choco-theme dependencies
All of the dependencies that are use by choco-theme have been upgraded
to the latest versions where possible. These mostly include development
build tools used by gulp and also yarn itself.

One big change was the switch from node-sass to dart-sass. This was
done because node-sass is now depreciated and was reporting security
vulnerabilities. With this change, slight modifications were needed in
some of the scss files in order to get the gulp builds working again.
This also requires each gulpfile.js to be updated in each repository to
use sass instead of node-sass.

With these updates, and the pinning of some packages in the resolutions
area in the package.json file, all security vulnerabilities are
resolved.