(#29) Add top alert maintenance banner
Updates choco-theme to 0.1.8 to include a top alert maintenance banner.

The gulpfile.js has been adjusted to allow the import of the TopAlertBanner
from choco-theme. There is also a local TopAlertBanner file that is still
in place in the repository itself, but this will be used specifically for alerts
to Boxstarter only.