(#59) Add container to docs layout
This adds a container to the content on the docs area
to limit how wide the content can go on larger screens.
This is inline with how docs.chocolatey.org and design.chocolatey.org
are set up. This helps users to be able to read content
better on larger screens.