(#61) Add AnchorJS anchors to list items
This adds anchors to list items that contain a `<strong>` or a `<code>`
node directly after an opening `<li>` tag. This is to make it easier to
navigate to these items, as often times this format is used more as a
"definition", or something we'd want to potentially link to.

No changes to current markdown is needed, as this intended to just
"work". All changes for this functionality live in choco-theme and will
be documented in the Choco Design System.