(#25) Re-add usage of Recipe
A previous attempt had been made to use the Chocolatey.Docs.Cake.Recipe
package, but it failed to work on GitHub Actions, so the changes were
reverted. I now believe that I know what the problem is, and it is to do
with the version of the Cake.Git addin.  For whatever reason, it fails
to work when running on Actions, but pinning to an earlier version of
the addin makes things work.  For now, going to put back in the reverted
changes, verify that things are still failing, and then get the required
fix added in.