(maint) Interim update of Bootstrapper script
Normally, direct changes to this bootstrapper.ps1 file would not be
made, however, in this instance, we are waiting for the next release of
Boxstarter 3.0.3, and while we are waiting, and to prevent other folks
running into issues with installing Boxstarter, a direct change is
being made.

This is related to this issue:
https://github.com/chocolatey/boxstarter/issues/548

The change here is so that we can get an immediate release of the
bootstrapper script, while we wait for the patch release of the
Boxstarter packages.

At that point, this Bootstrapper script will be updated with the new
package version, which will be 3.0.3.